Train Network in a Parallel process in NodeJS environment
Mimic WebWorker interface with Child Processes to allow parallel training
on NodeJS environment, and makes parallel training testable with mocha in
a terminal.

To accomplish this, it makes the method `Network::worker()` and
`Network::getWorkerSharedFunctions()` to generate a literal object, to
receive stringified methods, and giving meaning to this keyword.